add feature search and alert

+++ src/components/productDetails.jsx
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
+import Swal from 'sweetalert2'
+
 import ModalProduct from './modals/modalProduct'
 import {connect} from 'react-redux'
 import { getProductDetails, getProductsByBranchs, deleteProduct } from '../publics/redux/actions/products';
 import { getCategoryDetail } from '../publics/redux/actions/categories';
+
 
 class ProductDetails extends Component {
     state = {
@@ -32,7 +35,33 @@ class ProductDetails extends Component {
     }
 
     deleteProduct = async id => {
-        await this.props.dispatch(deleteProduct(id, this.state.category.name))
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            type: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+        }).then((result) => {
+            if (result.value) {
+                const Toast = Swal.mixin({
+                    toast: true,
+                    position: 'top-end',
+                    showConfirmButton: false,
+                    timer: 3000
+                })
+
+                Toast.fire({
+                    type: 'success',
+                    title: `Product ${this.state.product.name} has been deleted`
+                })
+                
+                setTimeout(()=>{
+                    this.props.dispatch(deleteProduct(id, this.state.category.name))
+                },1000)
+            }
+        })
     }
 
     render() { 
@@ -46,11 +75,16 @@ class ProductDetails extends Component {
                 <div className="col-md-8">
                     <div className="row">
                         <div className="col-md-8">
-                            <h4 style={{ fontWeight: 'bold' }}>{name}</h4>
+                            <h3 style={{ fontWeight: 'bold' }}>{name}</h3>
                         </div>
                         <div className="col-md-4 text-right">
                             <ModalProduct action="Edit" class="btn btn-secondary btn-sm mr-1"/>
                             <button className="btn btn-danger btn-sm ml-1" onClick={()=>this.deleteProduct(id)}>Delete</button>
+                        </div>
+                    </div>
+                    <div className="row">
+                        <div className="col-md-12">
+                            <h6 style={{color: '#F5D372'}}>{ this.state.category.name }</h6>
                         </div>
                     </div>
                     <div className="row pt-3">
